feat: support `--author` flag for git-shortlog
This patch adds support for `--author`, which passes the values straight
to `git-shortlog` [[1]].

Sometimes, it is useful to filter the shortlog by authors. In the past,
people might do this by piping the output of `git-stats` to `grep` or
similar. However, with the introduction of the Totals line, that line
becomes inaccurate.

This way, they can still get an accurate Total by filtering with `git`
directly. This can be useful, for example, for filtering out automated
commits (e.g., dependabot) or non-company contributors.

[1]: https://git-scm.com/docs/git-shortlog