fix: use `log`, not `shortlog` for proper `--author` support
This patch switches from using `git shortlog` to `git log` when
collecting the set of authors to compute stats on.

There is a bug with `git` itself when providing the `--author` flag.
Specifically, the behavior of `git log` and `git shortlog` behave differently
when dealing with author filters that also are affected by `.mailmap`.
`git log` appears to apply the filter AFTER applying `.mailmap`, whereas
`git shortlog` appears to do it BEFORE.

In practice, this means we often miss commits when people are filtering
with `--author` unintentionally.

For example, in a proprietary repository, here is the difference:

    ❯ git log --author="@numbersstation.ai" --format='%aN' | sort | uniq -c | sort -nr
       2442 Luke Hsiao
       1957 Nicholas Chiang
       1193 Sen Wu
       1089 Travis Chambers
       1016 Xiao Ling
        982 Vivek Raghuram
        939 Tywen Kelly
        744 Peng Xu
        647 Ines Chami
        591 Maureen Daum
        577 AJ Cloutier
        555 Laurel Orr
        486 Chris Lin
        386 David Kucher
         86 Kenny Daniel
         84 Rey Barceló
         35 Arun Khanchandani
         26 Sam Eklund
         20 Chao Li
          6 Long Le
          1 Satyajit Kumar
          1 Chris Aberger

    ❯ git shortlog --author="@numbersstation.ai" -sn
       1736  Luke Hsiao
       1083  Travis Chambers
        973  Xiao Ling
        848  Tywen Kelly
        702  Peng Xu
        591  Maureen Daum
        572  AJ Cloutier
        529  Vivek Raghuram
        486  Chris Lin
        482  Sen Wu
        421  Laurel Orr
        386  David Kucher
        381  Nicholas Chiang
         83  Rey Barceló
         35  Arun Khanchandani
         26  Sam Eklund
         20  Chao Li
          6  Long Le
          1  Chris Aberger
          1  Satyajit Kumar

Note that both the count and the list itself differ (e.g., Ines Chami
shows for `log`). The behavior of `git log` is the intended behavior.